[58] reader: compile ok
