+++ include/utils/types/error.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace error {
+     enum AppError {
+        kIOUnknown,
+    };
+} // namespace error
