+++ src/CMakeLists.txt
@@ -21,8 +21,8 @@ add_library(webserv_lib STATIC
     config/context/document_root.cpp
 	config/validator.cpp
 	utils/logger.cpp
-	io/read/reader_fd.cpp
-	io/read/buffer.cpp
+	io/input/read/buffer.cpp
+	io/input/reader/fd.cpp
 )
 
 # includeディレクトリの設定（ターゲットベース）
@@ -31,7 +31,10 @@ target_include_directories(webserv_lib PUBLIC
 	${PROJECT_SOURCE_DIR}/include/config
 	${PROJECT_SOURCE_DIR}/include/config/context
 	${PROJECT_SOURCE_DIR}/include/utils
+	${PROJECT_SOURCE_DIR}/include/utils/types
 	${PROJECT_SOURCE_DIR}/include/io
+	${PROJECT_SOURCE_DIR}/include/io/input/read
+	${PROJECT_SOURCE_DIR}/include/io/input/reader
 )
 
 # 実行ファイルの生成
